pyproject.toml
[project]
name = "imageanalysis"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "open3d (>=0.19.0,<0.20.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
